requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==24.4.2
Automat==24.8.1
backoff==2.2.1
bcrypt==4.2.1
boto==2.49.0
botocore==1.29.43
build==1.2.2.post1
cachetools==5.5.0
certifi==2021.10.8
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==2.1.1
chroma-hnswlib==0.7.6
chromadb==0.6.2
click==8.0.3
coloredlogs==15.0.1
constantly==23.10.4
contourpy==1.3.0
cryptography==44.0.2
cycler==0.11.0
daphne==4.1.2
Deprecated==1.2.15
distro==1.9.0
dj-database-url==0.5.0
Django==4.2.19
django-cors-headers==4.7.0
django-storages==1.13.2
django-storages-redux==1.3.3
djangorestframework==3.15.2
drf-yasg==1.21.7
durationpy==0.9
exceptiongroup==1.1.0
fastapi==0.115.6
filelock==3.16.1
flatbuffers==24.12.23
fonttools==4.55.3
frozenlist==1.3.3
fsspec==2024.12.0
gensim==3.8.0
git-filter-repo==2.47.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.27.1
humanfriendly==10.0
hyperlink==21.0.0
idna==3.3
importlib_metadata==8.5.0
importlib_resources==6.5.2
incremental==24.7.2
inflection==0.5.1
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.4
kubernetes==31.0.0
markdown-it-py==3.0.0
matplotlib==3.7.2
mdurl==0.1.2
mmh3==5.0.1
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
nest-asyncio==1.6.0
nltk==3.6.5
numpy==1.25.2
oauthlib==3.2.2
onnxruntime==1.19.2
openai==1.59.7
openai-cost-logger==0.4.1
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-instrumentation==0.50b0
opentelemetry-instrumentation-asgi==0.50b0
opentelemetry-instrumentation-fastapi==0.50b0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
opentelemetry-util-http==0.50b0
orjson==3.10.13
outcome==1.2.0
overrides==7.7.0
packaging==23.0
pandas==2.0.3
patsy==0.5.3
Pillow==8.4.0
posthog==3.7.5
protobuf==5.29.2
psycopg2-binary==2.9.5
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.21
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.19.1
pyOpenSSL==25.0.0
pyparsing==3.0.6
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2021.3
PyYAML==6.0.2
referencing==0.36.2
regex==2021.11.10
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.6.0
scikit-learn==1.3.0
scikit-posthocs==0.7.0
scipy==1.11.1
seaborn==0.12.2
selenium==4.8.2
service-identity==24.2.0
shellingham==1.5.4
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.3.0
sortedcontainers==2.4.0
spark-ai-python==0.4.5
sqlparse==0.4.3
starlette==0.41.3
statsmodels==0.13.5
style==1.1.0
swagger-spec-validator==3.0.4
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.0.0
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
trio==0.22.0
trio-websocket==0.9.2
trueskill==0.4.5
Twisted==24.11.0
txaio==23.1.1
typer==0.15.1
typing_extensions==4.11.0
tzdata==2024.2
update==0.0.1
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.3
websocket-client==1.8.0
websockets==14.1
wrapt==1.17.0
wsproto==1.2.0
yarl==1.8.2
yellowbrick==1.5
zipp==3.21.0
zope.interface==7.2
